Move common stackStates to its own file
some creep scripts contained states that other creeps also used, for easier finding of said states moved them all to a commonTasks file
